LIBFCREPO-1615. Created new plastron-jobs distribution package.
includes:
* plastron.jobs from plastron-repo
* ItemLog classes (in plastron.jobs.logs package) from plastron-utils

https://umd-dit.atlassian.net/browse/LIBFCREPO-1615
