+++ plastron-jobs/README.md
@@ -1,0 +1,10 @@
+# plastron-jobs
+
+Batch operation structures
+
+## Job Types
+
+* Export
+* [Import](docs/import.md)
+* Publication
+* Update
