+++ plastron-jobs/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "plastron-cli"
-description = "Plastron command line client"
+name = "plastron-jobs"
+description = "Batch job processing"
 authors = [
     { name='University of Maryland Libraries', email='lib-ssdr@umd.edu' },
     { name='Josh Westgard', email='westgard@umd.edu' },
@@ -13,17 +13,19 @@ authors = [
 readme = "README.md"
 requires-python = ">= 3.8"
 dependencies = [
-    "BeautifulSoup4",
-    "lxml",
-    "Pillow",
-    "plastron-jobs",
+    "bagit",
+    "beautifulsoup4",
+    "paramiko",
+    "plastron-client",
+    "plastron-messaging",
     "plastron-models",
     "plastron-repo",
     "plastron-utils",
     "pyparsing",
-    "pysolr",
     "PyYAML",
     "rdflib",
+    "requests",
+    "URLObject",
 ]
 dynamic = ["version"]
 
@@ -31,10 +33,8 @@ dynamic = ["version"]
 test = [
     "pytest",
     "pytest-cov",
+    "pytest-datadir",
 ]
-
-[project.scripts]
-plastron = "plastron.cli:main"
 
 [build-system]
 requires = ["setuptools>=66.1.0"]
@@ -42,6 +42,3 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools.dynamic]
 version = { "file" = "VERSION" }
-
-[tool.pytest.ini_options]
-markers = ['jobs_dir']
