+++ plastron-jobs/tests/jobs/test_logs/item_log.csv
@@ -1,0 +1,2 @@
+id,title
+foo,"The Adventures of Foo"
