+++ plastron-jobs/tests/jobs/test_publication_job.py
